+++ src/OFS/DTMLMethod.py
@@ -31,7 +31,6 @@ from OFS import bbb
 from OFS.Cache import Cacheable
 from OFS.role import RoleManager
 from OFS.SimpleItem import Item_w__name__
-from OFS.SimpleItem import PathReprProvider
 from six import binary_type
 from six import PY2
 from six import PY3
@@ -55,7 +54,6 @@ class Code(object):
 
 
 class DTMLMethod(
-    PathReprProvider,
     RestrictedDTML,
     HTML,
     Implicit,