+++ src/OFS/Folder.py
@@ -25,7 +25,6 @@ from OFS.ObjectManager import ObjectManager
 from OFS.PropertyManager import PropertyManager
 from OFS.role import RoleManager
 from OFS.SimpleItem import Item
-from OFS.SimpleItem import PathReprProvider
 from zope.interface import implementer
 
 
@@ -58,7 +57,6 @@ def manage_addFolder(
 
 @implementer(IFolder)
 class Folder(
-    PathReprProvider,
     ObjectManager,
     PropertyManager,
     RoleManager,